Switch adaptive to PID when in hover
Changed adaptive controller settings so that when the airplane gets up
to hover, the adaptive controller hands off to the PIDs so that the
adaptive doesnt saturate and go crazy.  Also changed the adaptive stuff
so that its a direct pass through to the control surfaces from u_p
instead of going through PIDs first.