Fixed aileron FFG offset problem but then disabled it
Forgot to only include aileron FF offset value when hover_flag was true
so it was causing right roll problems in all flight modes.  Fixed that
problem and did more tests but then disabled the FFG all together and
realized the stock PIDs were working better in the first place.